pyproject.toml
[project]
name = "tailorfish"
version = "0.1.0"
description = "An ITS-style chess training engine that turns your games into a personalized study plan."
readme = "README.md"
authors = [
    { name = "Al-Baraa El-Hag", email = "a.elhag01@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "python-chess>=1.999",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipython>=9.8.0",
    "mypy>=1.19.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.9",
]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src"]
extend-exclude = ["build", "dist", ".venv"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = ["E501"]  # let line-length handle it

[tool.mypy]
python_version = "3.12"
mypy_path = "src"
packages = ["tailorfish"]
strict = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
